src/backend/requirements.txt
fastapi
uvicorn[standard]
python-chess
